Treesame-Commit-Id: 303d7b4f6a8c9c4144decb4dfbdc97eb3cba0e0f

+++ db/migrate/20210901192734_add_replica_identity_for_pace_plan_module_items.rb
@@ -21,11 +21,15 @@ class AddReplicaIdentityForPacePlanModuleItems < ActiveRecord::Migration[6.0]
   disable_ddl_transaction!
 
   def up
+    return unless defined?(PacePlanModuleItem)
+
     add_replica_identity "PacePlanModuleItem", :root_account_id, 0
     remove_index :pace_plan_module_items, column: :root_account_id, if_exists: true
   end
 
   def down
+    return unless defined?(PacePlanModuleItem)
+
     add_index :pace_plan_module_items, :root_account_id, algorithm: :concurrently, if_not_exists: true
     remove_replica_identity "PacePlanModuleItem"
   end
